change location of app class to another file
