+++ Gemfile
@@ -3,3 +3,4 @@ source 'https://rubygems.org'
 git_source(:github) { |repo_name| "https://github.com/#{repo_name}" }
 
 # gem "rails"
+gem 'rubocop', '>= 1.0', '< 2.0'
